Move models-specific code to model subdirs
