+++ setup.cfg
@@ -60,12 +60,12 @@ zip_safe = False
 
 [options.entry_points]
 ewatercycle.models =
-    Hype = ewatercycle.models.hype:Hype
-    Lisflood = ewatercycle.models.lisflood:Lisflood
-    Wflow = ewatercycle.models.wflow:Wflow
-    PCRGlobWB = ewatercycle.models.pcrglobwb:PCRGlobWB
-    MarrmotM01 = ewatercycle.models.marrmot:MarrmotM01
-    MarrmotM14 = ewatercycle.models.marrmot:MarrmotM14
+    Hype = ewatercycle.plugins.hype.hype:Hype
+    Lisflood = ewatercycle.plugins.lisflood.lisflood:Lisflood
+    Wflow = ewatercycle.plugins.wflow.wflow:Wflow
+    PCRGlobWB = ewatercycle.plugins.pcrglobwb.pcrglobwb:PCRGlobWB
+    MarrmotM01 = ewatercycle.plugins.marrmot.marrmot:MarrmotM01
+    MarrmotM14 = ewatercycle.plugins.marrmot.marrmot:MarrmotM14
 
 [options.packages.find]
 where = src
