+++ src/ewatercycle/plugins/hype/hype.py
@@ -12,7 +12,7 @@ from dateutil.tz import UTC
 
 from ewatercycle import CFG
 from ewatercycle.container import VersionImages, start_container
-from ewatercycle.forcing._hype import HypeForcing
+from ewatercycle.plugins.hype.forcing import HypeForcing
 from ewatercycle.models.abstract import AbstractModel
 from ewatercycle.parameter_sets import ParameterSet
 from ewatercycle.util import geographical_distances, get_time, to_absolute_path
