+++ src/ewatercycle/plugins/lisflood/lisvap.py
@@ -29,8 +29,8 @@ from ewatercycle import CFG
 from ewatercycle.container import ContainerEngine
 from ewatercycle.parametersetdb.config import XmlConfig
 
-from ..config._lisflood_versions import version_images
-from ..util import get_time
+from ewatercycle.config._lisflood_versions import version_images
+from ewatercycle.util import get_time
 
 
 def lisvap(
