+++ src/ewatercycle/plugins/marrmot/forcing.py
@@ -5,9 +5,9 @@ from typing import Optional
 
 from esmvalcore.experimental import get_recipe
 
-from ..util import get_time, to_absolute_path
-from ._default import DefaultForcing, _session
-from .datasets import DATASETS
+from ewatercycle.util import get_time, to_absolute_path
+from ewatercycle.forcing._default import DefaultForcing, _session
+from ewatercycle.forcing.datasets import DATASETS
 
 
 class MarrmotForcing(DefaultForcing):
