+++ tests/models/test_hype.py
@@ -10,7 +10,7 @@ from grpc4bmi.bmi_client_apptainer import BmiClientApptainer
 
 from ewatercycle import CFG
 from ewatercycle.forcing import load_foreign
-from ewatercycle.models.hype import Hype, _set_code_in_cfg
+from ewatercycle.plugins.hype.hype import Hype, _set_code_in_cfg
 from ewatercycle.parameter_sets import ParameterSet
 from tests.models.fake_models import FailingModel
